Revert the move of jvm dependency analyzer to its own task.
JVM compilation in a given chunk must know whether fatal
missing dependencies are present, otherwise it will write
artifacts to the cache.  Even if the compile goal later
fails because of the missing dep detection task, the artifacts
have been written, and subsequent runs will have cache hits
which are not checked by missing dep detection and will
end in a successful run even though fatal missing deps are
present.

RB2721 was also reverted because it was a direct followup
that prevents RB2668 from being reverted cleanly.

This reverts RB 2721
Revert "Removed scan workunit; mapping workunits now debug"
This reverts commit ded93c1af9bce736550d1b0637148ab9125279a3.

This reverts RB 2668
Revert "Refactor JvmDependencyAnalyzer into a task"
This reverts commit a3c316a7d4ed9a9bdf1606b822a948be58db5eb6.
