+++ src/python/pants/backend/jvm/register.py
@@ -39,7 +39,6 @@ from pants.backend.jvm.tasks.junit_run import JUnitRun
 from pants.backend.jvm.tasks.jvm_compile.java.apt_compile import AptCompile
 from pants.backend.jvm.tasks.jvm_compile.java.java_compile import JmakeCompile
 from pants.backend.jvm.tasks.jvm_compile.scala.zinc_compile import ZincCompile
-from pants.backend.jvm.tasks.jvm_dependency_check import JvmDependencyCheck
 from pants.backend.jvm.tasks.jvm_platform_analysis import JvmPlatformExplain, JvmPlatformValidate
 from pants.backend.jvm.tasks.jvm_run import JvmRun
 from pants.backend.jvm.tasks.nailgun_task import NailgunKillall
@@ -123,7 +122,7 @@ def register_goals():
   # Compilation.
   jvm_compile = GroupTask.named(
       'jvm-compilers',
-      product_type=['classes_by_target', 'classes_by_source', 'resources_by_target', 'actual_source_deps'],
+      product_type=['classes_by_target', 'classes_by_source'],
       flag_namespace=['compile'])
 
   # It's important we add AptCompile before other java-compiling tasks since the first selector wins,
@@ -133,8 +132,6 @@ def register_goals():
   jvm_compile.add_member(ZincCompile)
 
   task(name='jvm', action=jvm_compile).install('compile').with_description('Compile source code.')
-  task(name='jvm-dep-check', action=JvmDependencyCheck).install('compile').with_description(
-      'Check that used dependencies have been requested.')
 
   # Generate documentation.
   task(name='javadoc', action=JavadocGen).install('doc').with_description('Create documentation.')
