+++ src/python/pants/backend/jvm/tasks/BUILD
@@ -19,7 +19,6 @@ python_library(
     ':javadoc_gen',
     ':junit_run',
     ':jvm_binary_task',
-    ':jvm_dependency_check',
     ':jvm_platform_analysis',
     ':jvm_run',
     ':jvm_task',
@@ -288,22 +287,6 @@ python_library(
     ':jar_task',
     '3rdparty/python/twitter/commons:twitter.common.collections',
     'src/python/pants/backend/jvm/targets:jvm',
-  ],
-)
-
-python_library(
-  name = 'jvm_dependency_check',
-  sources = ['jvm_dependency_check.py'],
-  dependencies = [
-    '3rdparty/python/twitter/commons:twitter.common.collections',
-    'src/python/pants/backend/core/tasks:task',
-    'src/python/pants/base:build_environment',
-    'src/python/pants/base:build_graph',
-    'src/python/pants/base:exceptions',
-    'src/python/pants/backend/jvm/tasks:ivy_task_mixin',
-    'src/python/pants/backend/jvm/targets:jvm',
-    'src/python/pants/backend/jvm/targets:scala',
-    'src/python/pants/util:memo',
   ],
 )
