+++ src/python/pants/backend/jvm/tasks/jvm_compile/jvm_compile.py
@@ -6,7 +6,6 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
                         unicode_literals, with_statement)
 
 import itertools
-import os
 import sys
 from collections import defaultdict
 
@@ -15,6 +14,7 @@ from pants.backend.jvm.subsystems.jvm_platform import JvmPlatform
 from pants.backend.jvm.tasks.jvm_compile.jvm_compile_global_strategy import JvmCompileGlobalStrategy
 from pants.backend.jvm.tasks.jvm_compile.jvm_compile_isolated_strategy import \
   JvmCompileIsolatedStrategy
+from pants.backend.jvm.tasks.jvm_compile.jvm_dependency_analyzer import JvmDependencyAnalyzer
 from pants.backend.jvm.tasks.nailgun_task import NailgunTaskBase
 from pants.base.fingerprint_strategy import TaskIdentityFingerprintStrategy
 from pants.base.workunit import WorkUnitLabel
@@ -79,12 +79,14 @@ class JvmCompile(NailgunTaskBase, GroupMember):
 
   @classmethod
   def product_types(cls):
-    raise TaskError('Expected to be installed in GroupTask, which has its own '
-                    'product_types implementation.')
+    return ['classes_by_target', 'classes_by_source', 'resources_by_target']
 
   @classmethod
   def prepare(cls, options, round_manager):
     super(JvmCompile, cls).prepare(options, round_manager)
+
+    # This task uses JvmDependencyAnalyzer as a helper, get its product needs
+    JvmDependencyAnalyzer.prepare(options, round_manager)
 
     round_manager.require_data('compile_classpath')
     round_manager.require_data('ivy_resolve_symlink_map')
@@ -225,8 +227,6 @@ class JvmCompile(NailgunTaskBase, GroupMember):
                                           self.create_analysis_tools(),
                                           self._name,
                                           self.select_source)
-    # Maps CompileContext --> dict of upstream class to paths.
-    self._upstream_class_to_paths = {}
 
   def _fingerprint_strategy(self):
     return TaskIdentityFingerprintStrategy(self)
@@ -340,17 +340,14 @@ class JvmCompile(NailgunTaskBase, GroupMember):
     # uses it internally.
     self.context.products.safe_create_data('resources_by_target', make_products)
 
-    # JvmDependencyCheck uses classes_by_target
+    # JvmDependencyAnalyzer uses classes_by_target within this run
     self.context.products.safe_create_data('classes_by_target', make_products)
-
-    self.context.products.safe_create_data('actual_source_deps', dict)
 
   def _register_vts(self, compile_contexts):
     classes_by_source = self.context.products.get_data('classes_by_source')
     classes_by_target = self.context.products.get_data('classes_by_target')
     compile_classpath = self.context.products.get_data('compile_classpath')
     resources_by_target = self.context.products.get_data('resources_by_target')
-    actual_source_deps = self.context.products.get_data('actual_source_deps')
 
     # Register class products (and resources generated by annotation processors.)
     computed_classes_by_source_by_context = self._strategy.compute_classes_by_source(
@@ -390,8 +387,8 @@ class JvmCompile(NailgunTaskBase, GroupMember):
         )
         add_products_by_target(unclaimed_classes)
 
+    # Register resource products.
     for compile_context in compile_contexts:
-      # Register resource products.
       extra_resources = self.extra_products(compile_context.target)
       # Add to resources_by_target (if it was requested).
       if resources_by_target is not None:
@@ -403,8 +400,3 @@ class JvmCompile(NailgunTaskBase, GroupMember):
       # are not available during compilation. https://github.com/pantsbuild/pants/issues/206
       entries = [(conf, root) for conf in self._confs for root, _ in extra_resources]
       compile_classpath.add_for_target(compile_context.target, entries)
-
-      if self.context.products.is_required_data('actual_source_deps'):
-        classpath = [path for (conf, path) in entries if conf in self._confs]
-        actual_source_deps[compile_context.target] = self._strategy.parse_deps(classpath,
-                                                                               compile_context)
