+++ src/python/pants/backend/jvm/tasks/jvm_compile/jvm_compile_isolated_strategy.py
@@ -369,9 +369,3 @@ class JvmCompileIsolatedStrategy(JvmCompileStrategy):
           update_artifact_cache_work
       ]
       self.context.submit_background_work_chain(work_chain, parent_workunit_name='cache')
-
-  def parse_deps(self, classpath, compile_context):
-    # We intentionally pass in an empty string for classes_dir in order to receive
-    # relative paths, because we currently don't know (and don't need to know) what target
-    # provided the file.
-    return self._analysis_parser.parse_deps_from_path(compile_context.analysis_file, dict, '')
