+++ src/python/pants/backend/jvm/tasks/jvm_compile/jvm_compile_strategy.py
@@ -225,11 +225,3 @@ class JvmCompileStrategy(object):
         lambda: safe_rmtree(analysis_tmpdir))
     safe_mkdir(analysis_tmpdir)
     return analysis_tmpdir
-
-  @abstractmethod
-  def parse_deps(self, classpath, compile_context):
-    """Parses the actual source dependencies of compile_context.target given a classpath.
-
-    The returned source dependencies may be either absolute filepaths, or relative filepaths.
-    """
-    pass
